# add elf parsing utility function