This PR introduces some basic ELF parsing logic, which is useful in various
places.

Example uses:

- Retrieve all ELF executables and shared libs in a prefix
- Get the offset of the rpath string in a file so it can be modified in-place
- Get better info about why a file has no rpath (e.g. not an elf file,
  malformed elf, no dynamic section, missing tag in dynamic array, ...), this
  is more fine-grained than in patchelf.

It's very basic, and it's really about parsing, since we don't want to reimplement
the advanced dynamic section growth logic that patchelf offers.

Current API is basically:

```python
import spack.util.elf as elf

try:
    with open("my_binary", "br") as f:
        parsed = elf.parse_elf(f)
except elf.ElfParsingError:
    # handle invalid elf file

# booleans
parsed.has_pt_interp
parsed.has_pt_dynamic
parsed.has_rpath
parsed.has_needed
parsed.has_soname
parsed.is_runpath

# byte strings
parsed.pt_interp_str
parsed.dt_soname_str
parsed.dt_rpath_str
parsed.dt_needed_strs  # list of byte strings

# more low-level offsets into the file, or the strtab associated with the dynamic section (useful for in-place updates)
```

Another advantage is that we get a ~850x speedup for retrieving rpaths.

```python
In [1]: spack.relocate._elf_rpaths_for("libclang.so.15.0.0") == spack.util.elf.get_rpaths("libclang.so.15.0.0")
Out[1]: True

In [2]: %timeit spack.relocate._elf_rpaths_for("libclang.so.15.0.0")
42.8 ms ± 448 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

In [3]: %timeit spack.util.elf.get_rpaths("libclang.so.15.0.0")
50.4 µs ± 435 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```