@trws 2022-11-01
I like this, a lot, and the test changes are really welcome.  Had a couple of questions but I don't see anything that would make me want to keep this out.